+++ src/Runner.php
@@ -11,7 +11,7 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 
 class Runner {
 
-    const VERSION = '0.3.3';
+    const VERSION = '0.3.4';
     const ROBOCLASS = 'RoboFile';
     const ROBOFILE = 'RoboFile.php';